# packages/server/src/schema/zod/applicationConfigSchema.ts
/**
 * Generated by @kubb/plugin-zod - Do not edit by hand
 * @see https://github.com/kubb-labs/kubb/tree/main/packages/plugin-zod
 */

import { z } from 'zod'

export const applicationConfigSchema = z.object({
  displayName: z.string().describe("Friendly name of the application config, eg; 'dev'"),
  clientId: z.string().describe('Client ID of the application'),
  scopes: z.array(z.string()).describe('Scopes of the application; openid, profile, email, offline_access, ...'),
  domain: z.string().describe('The DNS on which the application is hosted'),
  redirectUrl: z.string().describe('The redirect URL of the application'),
  backendUrl: z.string().describe('The backend URL of the application').optional(),
})

export type ApplicationConfigSchema = z.infer<typeof applicationConfigSchema>