# packages/server/src/schema/zod/applicationSchema.ts
/**
 * Generated by @kubb/plugin-zod - Do not edit by hand
 * @see https://github.com/kubb-labs/kubb/tree/main/packages/plugin-zod
 */

import { applicationConfigSchema } from './applicationConfigSchema.ts'
import { z } from 'zod'

export const applicationSchema = z.object({
  id: z.string().describe('Object ID'),
  displayName: z.string().describe('Name of the application to display in the UI'),
  roles: z.array(z.string()).describe('The roles that are available for this application'),
  description: z.string().describe('Description of the application').optional(),
  slug: z.string().describe('Shortcut to be used in role assignment `tsz:admin`'),
  createdOn: z.string().describe('The date & time the application was created'),
  updatedOn: z.string().describe('The date & time the application last updated'),
  config: z.array(z.lazy(() => applicationConfigSchema)).describe('The configuration of the application, used by the TokenHandler'),
})

export type ApplicationSchema = z.infer<typeof applicationSchema>