# packages/server/src/schema/zod/badRequestErrorSchema.ts
/**
 * Generated by @kubb/plugin-zod - Do not edit by hand
 * @see https://github.com/kubb-labs/kubb/tree/main/packages/plugin-zod
 */

import { errorDetailSchema } from './errorDetailSchema.ts'
import { z } from 'zod'

export const badRequestErrorSchema = z.object({
  status: z.number().int(),
  title: z.string(),
  errors: z.array(z.lazy(() => errorDetailSchema)),
})

export type BadRequestErrorSchema = z.infer<typeof badRequestErrorSchema>