# packages/server/src/schema/zod/clientSecretSchema.ts
/**
 * Generated by @kubb/plugin-zod - Do not edit by hand
 * @see https://github.com/kubb-labs/kubb/tree/main/packages/plugin-zod
 */

import { z } from 'zod'

export const clientSecretSchema = z.object({
  id: z.string(),
  description: z.string().describe('The secret name'),
  value: z.string().describe('The secret value, masked when read'),
  expiresAt: z.string().date().describe('The date & time the secret is expiring'),
})

export type ClientSecretSchema = z.infer<typeof clientSecretSchema>