# packages/server/src/schema/zod/errorSchema.ts
/**
 * Generated by @kubb/plugin-zod - Do not edit by hand
 * @see https://github.com/kubb-labs/kubb/tree/main/packages/plugin-zod
 */

import { z } from 'zod'

export const errorSchema = z.object({
  status: z.number().int(),
  title: z.string(),
})

export type ErrorSchema = z.infer<typeof errorSchema>