# packages/server/src/schema/zod/notFoundErrorSchema.ts
/**
 * Generated by @kubb/plugin-zod - Do not edit by hand
 * @see https://github.com/kubb-labs/kubb/tree/main/packages/plugin-zod
 */

import { z } from 'zod'

export const notFoundErrorSchema = z.object({
  status: z.number().int(),
  title: z.string(),
})

export type NotFoundErrorSchema = z.infer<typeof notFoundErrorSchema>